+++ code/alma/implement/ASDMBinaries/SDMDataObjectWriter.h
@@ -749,7 +749,7 @@ namespace asdmbinaries {
     void outputln (const string& s);
     void output   (const float* data, uint32_t numData);
     void outputln (const float* data, uint32_t numData);
-    void outputln (const int64_t* data, uint32_t numData);
+    void outputln (const long long* data, uint32_t numData);
 
     template <class T> void output(const vector<T>& data) {
       numBytes_ += data.size()*sizeof(T);
